requirements.txt
arxiv==2.1.3
llama-index-vector-stores-faiss
faiss-cpu
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface